Move to_abstract and type_to_abstract to myia.to_abstract.py
