+++ myia/abstract/__init__.py
@@ -6,4 +6,5 @@ from .infer import *  # noqa
 from .loop import *  # noqa
 from .macro import *  # noqa
 from .ref import *  # noqa
+from .to_abstract import *  # noqa
 from .utils import *  # noqa
