+++ myia/abstract/data.py
@@ -607,7 +607,7 @@ class AbstractClassBase(AbstractStructure):
         This uses the attribute types as defined by the user, rather than what
         is generated by the inferrer or other methods.
         """
-        from .utils import type_to_abstract
+        from .to_abstract import type_to_abstract
         return type_to_abstract(self.tag)
 
     def children(self):
